requirements.txt
PyJWT
bcrypt
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.11.4
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0